requirements.txt
Django==1.11.7
django-crispy-forms==1.7.0
djangorestframework==3.7.1
pytz==2017.3
